requirements-api.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
joblib
